requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86  # Required for ximgproc module
pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
werkzeug==3.1.3
gunicorn==23.0.0
numpy>=1.20.0
tqdm>=4.64.0  # Progress bar utility
scikit-image>=0.21.0  # For advanced image processing
matplotlib>=3.7.0  # For visualization if needed
psutil>=5.9.0  # For system monitoring
